renamed analysis dir to "nml"
