+++ scripts/nml/analyze_ifs.py
