+++ scripts/nml/analyze_namelist.py
